afwezigheid/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "afwezigheid"
version = "0.1.0"
description = "Absence management system"
requires-python = ">=3.8"
dependencies = [
    "flask>=2.3.0",
    "flask-login>=0.6.0",
    "flask-sqlalchemy>=3.0.0",
    "python-dotenv>=1.0.0",
    "mysql-connector-python>=8.0.0",
    "jinja2>=3.0.0",
    "pydantic[email]>=2.0.0",
    "flask-wtf>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
